requirements.txt
python_slugify==8.0.1
pickledb==0.9.2
openai==1.14.1
pypandoc==1.13
GitPython==3.1.43
